+++ src/navigation/RootStackNavigation.tsx
@@ -30,7 +30,7 @@ import ReceiveNavigation from '~/navigation/ReceiveNavigation'
 import RootStackParamList from '~/navigation/rootStackRoutes'
 import SendNavigation from '~/navigation/SendNavigation'
 import EditAddressScreen from '~/screens/Address/EditAddressScreen'
-import AddressDiscoveryScreen from '~/screens/AddressDiscovery'
+import AddressDiscoveryScreen from '~/screens/AddressDiscoveryScreen'
 import AddressQuickNavigationScreen from '~/screens/Addresses/AddressQuickNavigationScreen'
 import ContactScreen from '~/screens/Addresses/Contact/ContactScreen'
 import EditContactScreen from '~/screens/Addresses/Contact/EditContactScreen'
